Implement first version of the Ip control
This version includes a class for Ip control, ControlSolenoid, and a
class for scheduling the relevant entities for the solenoid,
SolenoidScheduler.

The former scheduler class has been split into a parent class
(TargetScheduler, in the target_scheduler module) and two children
classes for scheduling different targets. This commit includes
SolenoidScheduler, the child class of TargetScheduler for scheduling
the changes to apply to the solenoid.

These two classes haven't been tested, so there is no guarantee
whatsoever that it will not throw errors if executed as is.